Initial battle assets
Battle loads now, but it's very glitchy and doesn't accept player input so the monsters will pretty much always win and result in a game over. I think all of the 24-bit references in the battle mechanics code are working, just need to finish the ones in the battle graphics code.